# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QueryGenie API
  description: |
    Enterprise-grade API for QueryGenie - AI-powered database assistant with multi-role collaboration.
    
    ## Authentication
    All API endpoints require authentication via JWT tokens obtained through NextAuth.js.
    
    ## Rate Limiting
    - **Free Tier**: 100 requests per 15 minutes
    - **Pro Tier**: 1000 requests per 15 minutes  
    - **Enterprise**: Custom limits
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format.
    
  version: 1.0.0
  contact:
    name: QueryGenie API Support
    url: https://querygenie.com/support
    email: support@querygenie.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.querygenie.com
    description: Production server
  - url: https://staging-api.querygenie.com
    description: Staging server
security:
  - BearerAuth: []
  - SessionAuth: []

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      summary: Get workspace details
      description: Returns detailed information about a workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceDetails'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    patch:
      tags: [Workspaces]
      summary: Update workspace
      description: Updates workspace information (requires owner/admin role)
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    delete:
      tags: [Workspaces]
      summary: Delete workspace
      description: Permanently deletes a workspace (requires owner role)
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '204':
          description: Workspace deleted successfully
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /workspaces/{workspaceId}/queries:
    get:
      tags: [Queries]
      summary: List queries
      description: Returns queries in a workspace (history and saved)
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: type
          in: query
          schema:
            type: string
            enum: [all, saved, history]
            default: all
        - name: connectionId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of queries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Query'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Queries]
      summary: Execute or save query
      description: Executes a SQL query or saves it for later use
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteQueryRequest'
      responses:
        '200':
          description: Query execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResult'
        '202':
          description: Query execution started (for long-running queries)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token

  # Common Parameters
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        format: uuid
      example: 123e4567-e89b-12d3-a456-426614174000
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 8
    JobResponse:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        estimatedDuration:
          type: integer
          description: Estimated duration in seconds

    # User Types
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
        role:
          type: string
          enum: [admin, developer, analyst]
        emailVerified:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    WorkspaceRole:
      type: string
      enum: [owner, admin, editor, viewer]
      description: |
        - **owner**: Full access including deletion
        - **admin**: Manage connections, members, settings
        - **editor**: Create/execute queries, edit docs  
        - **viewer**: Read-only access
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
        ownerId:
          type: string
          format: uuid
        isActive:
          type: boolean
        memberCount:
          type: integer
        connectionCount:
          type: integer
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceDetails:
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          properties:
            owner:
              $ref: '#/components/schemas/User'
            settings:
              type: object
              additionalProperties: true
            members:
              type: array
              items:
                $ref: '#/components/schemas/WorkspaceMember'
            recentActivity:
              type: array
              items:
                $ref: '#/components/schemas/AuditLog'
...
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        avatar:
          type: string
          format: uri
        settings:
          type: object
          additionalProperties: true
    WorkspaceMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/User'
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        isActive:
          type: boolean
        joinedAt:
          type: string
          format: date-time
...
    ConnectionType:
      type: string
      enum: [postgres, mysql, mssql, sqlite]
    ConnectionMode:
      type: string
      enum: [read-only, read-write]
    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: '#/components/schemas/ConnectionType'
        mode:
          $ref: '#/components/schemas/ConnectionMode'
        host:
          type: string
          nullable: true
        port:
          type: integer
          nullable: true
        database:
          type: string
          nullable: true
        isActive:
          type: boolean
        lastConnectedAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
...
    QueryStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    Query:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        prompt:
          type: string
          nullable: true
          description: Original natural language prompt
        sqlQuery:
          type: string
        status:
          $ref: '#/components/schemas/QueryStatus'
        executionTime:
          type: integer
          nullable: true
          description: Execution time in milliseconds
        rowsAffected:
          type: integer
          nullable: true
        errorMessage:
          type: string
          nullable: true
        isShared:
          type: boolean
        isSaved:
          type: boolean
        tags:
          type: string
          nullable: true
          description: Comma-separated tags
        connection:
          $ref: '#/components/schemas/Connection'
        createdBy:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
...
    ExecuteQueryRequest:
      type: object
      required:
        - sqlQuery
        - connectionId
      properties:
        sqlQuery:
          type: string
          minLength: 1
        connectionId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        isSaved:
          type: boolean
          default: false
        tags:
          type: string
          description: Comma-separated tags
        limit:
          type: integer
          minimum: 1
          maximum: 10000
          default: 100
          description: Row limit for SELECT queries
    QueryResult:
      type: object
      properties:
        query:
          $ref: '#/components/schemas/Query'
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        columns:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
        rowCount:
          type: integer
        executionTime:
          type: integer
        hasMore:
          type: boolean
          description: True if there are more rows available
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          enum: [create, read, update, delete, login, logout, invite, remove, execute_query, generate_docs, export_data]
        resource:
          type: string
        resourceId:
          type: string
          format: uuid
          nullable: true
        user:
          $ref: '#/components/schemas/User'
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    # System Types
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem type
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Request trace ID for debugging
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
                  value:
                    description: The invalid value

  # Common Responses
  responses:
...
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: https://querygenie.com/errors/forbidden
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action
            instance: /api/workspaces/123/connections
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: https://querygenie.com/errors/not-found
            title: Not Found
            status: 404
            detail: The requested resource was not found
            instance: /api/workspaces/invalid-id
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            type: https://querygenie.com/errors/validation
            title: Validation Failed
            status: 400
            detail: One or more fields contain invalid values
            instance: /api/workspaces
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
            errors:
              - field: name
                code: required
                message: Name is required
              - field: slug
                code: pattern
                message: Slug must contain only lowercase letters, numbers, and hyphens
                value: "Invalid Slug!"
...